+++ src/main/kotlin/christmas/domain/promotion/WeekdayDessertDiscount.kt
@@ -1,21 +1,21 @@
-package christmas.configuration.promotion
+package christmas.domain.promotion
 
-import christmas.configuration.promotion.util.DayChecker.isWeekend
 import christmas.domain.money.Currency
 import christmas.domain.money.Money
 import christmas.domain.order.Orders
 import christmas.domain.product.Product
 import christmas.domain.product.ProductType
+import christmas.domain.promotion.util.DayChecker.isWeekday
 
-object WeekendMainDishDiscount : Promotion {
+object WeekdayDessertDiscount : Promotion {
     private val DISCOUNT_AMOUNT = Money.longValueOf(2023, Currency.KRW)
 
     override fun discountAmount(orders: Orders): Money {
         val dayOfWeek = orders.date.dayOfWeek
 
-        if (dayOfWeek.isWeekend()) {
-            val mainCount = countMainDishes(orders)
-            val totalDiscountMoney = DISCOUNT_AMOUNT.multiply(mainCount)
+        if (dayOfWeek.isWeekday()) {
+            val dessertCount = countDesserts(orders)
+            val totalDiscountMoney = DISCOUNT_AMOUNT.multiply(dessertCount)
             return totalDiscountMoney
         }
 
@@ -26,9 +26,9 @@ object WeekendMainDishDiscount : Promotion {
         return null
     }
 
-    private fun countMainDishes(orders: Orders): Int {
+    private fun countDesserts(orders: Orders): Int {
         return orders.orders
-            .filter { it.product.productType == ProductType.MAIN }
+            .filter { it.product.productType == ProductType.DESSERT }
             .sumOf { order -> order.amount }
     }
 }
