+++ src/main/kotlin/christmas/domain/promotion/util/DayChecker.kt
@@ -1,4 +1,4 @@
-package christmas.configuration.promotion.util
+package christmas.domain.promotion.util
 
 import java.time.DayOfWeek
 import java.time.DayOfWeek.FRIDAY
