+++ src/main/kotlin/christmas/service/PromotionService.kt
@@ -1,11 +1,17 @@
-package christmas.configuration.promotion
+package christmas.service
 
 import christmas.domain.money.Currency
 import christmas.domain.money.Money
 import christmas.domain.order.Orders
 import christmas.domain.product.Product
+import christmas.domain.promotion.ChampagneGiveaway
+import christmas.domain.promotion.ChristmasDdayDiscount
+import christmas.domain.promotion.Promotion
+import christmas.domain.promotion.StarDiscount
+import christmas.domain.promotion.WeekdayDessertDiscount
+import christmas.domain.promotion.WeekendMainDishDiscount
 
-data class Promotions(
+object PromotionService {
     private val promotions: List<Promotion> = listOf(
         ChampagneGiveaway,
         ChristmasDdayDiscount,
@@ -13,7 +19,7 @@ data class Promotions(
         WeekdayDessertDiscount,
         WeekendMainDishDiscount,
     )
-) {
+
     fun findAppliedPromotions(orders: Orders): List<Promotion> {
         return promotions.filter { promotion -> isPromoted(promotion, orders) }
     }
@@ -26,7 +32,7 @@ data class Promotions(
     }
 
     private fun isNotPromoted(discountAmount: Money, giveawayProduct: Product?): Boolean {
-        return discountAmount == Money.longValueOf(0, Currency.KRW)
+        return discountAmount == Money.Companion.longValueOf(0, Currency.KRW)
                 && giveawayProduct == null
     }
 }
